requirements-optional-packages.txt
cleopatra >=0.4.0
gdal >=3.7.0, <3.7.2
geopandas >=0.13.2
hpc-utils ==0.1.2
loguru >=0.7.0
numpy ==1.25.2
pandas >=2.0.0
pip >=23.2.1
pyproj >=3.6.0
PyYAML >=6.0.1
Shapely >=2.0.1
